pyproject.toml
[project]
name = "unmonitorr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.11",
    "coloredlogs>=15.0.1",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.0.1",
]
